Removed group management
Managing the puppetadmins group from this module turned out to cause nasty
dependencies to this module in places where they're not need or wanted. Removed
group management part of this module and created an explicit dependency to
private, yet easily re-implementable "localusers" module, which must handle the
creation of the appropriate puppetadmins group.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>